+++ packages/analytics/dashboard-renderer/src/types/renderer-types.ts
@@ -16,8 +16,25 @@ export interface DashboardRendererContext {
   refreshInterval?: number
   editable?: boolean
   showTileActions?: boolean
-  zoomable?: boolean
-  showTileZoomActions?: boolean
+}
+
+export type ZoomConfiguration = {
+  /**
+   * whether or not to allow zoom on a chart (i.e. click/drag/unclick to highlight a range)
+   */
+  enabled: boolean
+  /**
+   * whether the context menu should include "Zoom in"
+   */
+  showZoomInAction: boolean
+  /**
+   * whether or not the context menu should include "Explore"
+   */
+  showExploreAction: boolean
+  /**
+   * whether or not the context menu should include "Jump to requests"
+   */
+  showRequestsAction: boolean
 }
 
 export interface PdfExportOptions {
@@ -77,6 +94,7 @@ export interface ChartRendererProps<T> {
   headerDescription?: string
   requestsLink?: ExternalLink
   exploreLink?: ExternalLink
+  zoomConfiguration?: ZoomConfiguration
 }
 
 export interface ScatterRendererProps extends Omit<ChartRendererProps<ScatterChartOptions>, 'query'> {